+++ src/app/ikm-expo/page.tsx
@@ -1,20 +1,58 @@
 import { redirect } from "next/navigation";
+import { createClient } from "@/utils/supabase/server";
+import IkmExpoSuccessPage from "./_components/success";
+import IkmExpoExpiredPage from "./_components/expired";
 
 export default async function IkmExpoPage() {
-    const baseUrl =
-        process.env.NEXT_PUBLIC_BASE_URL || "https://bootcamp.exerciseftui.com";
-    const res = await fetch(`${baseUrl}/api/ikm-expo/claim`, {
-        method: "POST",
-    });
-    console.log("res", res);
-    const result = await res.json();
-    if (result.success) {
-        redirect("/ikm-expo/success");
-    } else if (result.reason === "expired") {
-        redirect("/ikm-expo/expired");
-    } else if (result.reason === "unauthenticated") {
+    // Create Supabase client for server-side authentication
+    const supabase = createClient();
+
+    const {
+        data: { user },
+        error: userError,
+    } = await (await supabase).auth.getUser();
+
+    // Check if user is authenticated
+    if (!user || userError) {
         redirect("/auth/login?returnTo=/ikm-expo");
-    } else {
-        redirect("/");
+    }
+
+    // Check if user already claim the campaign
+    const { data: userRecord } = await (await supabase)
+        .from("users")
+        .select("isFromIKMExpo")
+        .eq("user_id", user.id)
+        .single();
+
+    if (userRecord?.isFromIKMExpo) {
+        redirect("/dashboard");
+    }
+
+    // Check if campaign is still active
+    const CAMPAIGN_DEADLINE = new Date("2025-09-19T16:59:59Z");
+    const now = new Date();
+
+    if (now >= CAMPAIGN_DEADLINE) {
+        return <IkmExpoExpiredPage />;
+    }
+
+    // Update user's isFromIKMExpo status
+    try {
+        console.log(`Updating user ${user.id} -> isFromIKMExpo = true`);
+        const { error } = await (await supabase)
+            .from("users")
+            .update({ isFromIKMExpo: true })
+            .eq("user_id", user.id);
+
+        if (error) {
+            console.error("Database update failed:", error);
+            redirect("/dashboard");
+        }
+
+        console.log("Update succeeded. Showing success page.");
+        return <IkmExpoSuccessPage />;
+    } catch (error) {
+        console.error("Unexpected error:", error);
+        return <IkmExpoExpiredPage />;
     }
 }
